Object interaction
I think interaction with doors is similar with interaction with books on the ground. And since that would be an important aspect of the game. I tried to make that mechanism and failed... I am still debugging because the character is not picking up the book. Everything is in "Justin" scene.